ver 2.0

+++ swim_site/swim_site_backend/backend/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Child(models.Model):
     name = models.CharField(max_length=255, verbose_name='Фио ребёнка', unique=True)
     paid_training_count = models.IntegerField(default=0, verbose_name='Кол-во оплаченных занятий')
-    parent_chat_id = models.IntegerField(blank=True, null=True, verbose_name='Айди чата родителя')
+    parent_chat_id = models.BigIntegerField(blank=True, null=True, verbose_name='Айди чата родителя')
     last_balance_update = models.DateField(blank=True, null=True, verbose_name='Дата последнего пополнения')
 
     def __str__(self):
@@ -54,7 +54,7 @@ class Training(models.Model):
 
 class ChildId(models.Model):
     name = models.CharField(max_length=255, verbose_name='Имя ребенка', unique=True)
-    parent_chat_id = models.IntegerField(verbose_name='Нынешний айди', unique=True)
+    parent_chat_id = models.BigIntegerField(verbose_name='Нынешний айди', unique=True)
 
 
 class BalanceOperations(models.Model):
